Moved man for FastNetMon daemon from 8 to 1 as it's best preactice.
The command in /sbin or /usr/sbin are system administration commands; their manual pages
thus belong in section 8, not section 1.

Please check whether the command is actually useful to non-privileged user in which
case it should be moved to /bin or /usr/bin, or alternatively the manual page should
be moved to   section 8 instead, ie. /usr/share/man/man8.   Please refer to Bug#348864,
Bug#253011, and the hier(7) manual page for details.